Put more error check for shader resource view proxy
